pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "psipy"
version = "0.1.0"
description = "Un projet pour la physique symbolique et la résolution d'EDP."
authors = [
    {name = "Philippe Billet", email = "ton@email.com"},
]
dependencies = [
    "numpy==2.2.6",
    "sympy",
    "scipy",
    "matplotlib==3.10.6",
    "pandas",
    "jupyter",
    "ipython",
    "notebook",
    "tqdm",
    "requests",
    "flask",
    "scikit-learn",
    "scikit-image",
    "black",
    "flake8",
    "isort",
    "pylint",
    "ruff",
    "setuptools==78.1.1",
    "pytest==8.2.2",
    "pytest-cov==6.2.1",
    "sphinx",
    "numpydoc",
    "pydata-sphinx-theme",
    "myst-nb",
    "myst-parser",
    "xarray",
    "h5py",
    "netcdf4",
    "dask",
    "bokeh",
    "panel",
    "holoviews",
    "datashader",
    "streamlit",
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pyyaml",
    "pyodbc",
    "pillow",
    "opencv-python",
    "tifffile",
    "imageio",
    "pooch==1.8.2",
    "soundfile==0.13.1",
    "librosa==0.11.0",
    "pywavelets",
    "statsmodels",
    "beautifulsoup4",
    "lxml",
    "html5lib",
    "soupsieve",
    "pyopenssl",
    "cryptography",
    "paramiko",
    "pygithub",
    "gitpython",
    "pyzmq",
    "tornado",
    "jupyterlab",
    "widgetsnbextension",
    "ipywidgets",
    "nbconvert",
    "nbformat",
    "traitlets",
    "jupyter_client",
    "jupyter_console",
    "jupyter_core",
    "prometheus_client",
    "s3fs",
    "fsspec",
    "aiohttp",
    "aiobotocore",
    "botocore",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
dev = [
    "black>=23.0",
    "flake8>=6.0",
    "isort>=5.0",
    "pylint>=3.0",
    "ruff>=0.1.0",
]
